Move repo.go into new package.
Avoid circular dependency with charmhub, testing and testcharms.
