+++ state/export_test.go
@@ -40,6 +40,7 @@ import (
 	"github.com/juju/juju/state/storage"
 	"github.com/juju/juju/state/watcher"
 	"github.com/juju/juju/testcharms"
+	"github.com/juju/juju/testcharms/repo"
 	"github.com/juju/juju/version"
 )
 
@@ -230,7 +231,7 @@ func AddTestingCharmWithSeries(c *gc.C, st *State, name string, series string) *
 	return addCharm(c, st, series, testcharms.Repo.CharmDir(name))
 }
 
-func getCharmRepo(series string) *testcharms.CharmRepo {
+func getCharmRepo(series string) *repo.CharmRepo {
 	// ALl testing charms for state are under `quantal` except `kubernetes`.
 	if series == "kubernetes" {
 		return testcharms.RepoForSeries(series)
@@ -383,7 +384,7 @@ func addTestingApplication(c *gc.C, params addTestingApplicationParams) *Applica
 	return app
 }
 
-func addCustomCharmWithManifest(c *gc.C, st *State, repo *testcharms.CharmRepo, name, filename, content, series string, revision int, manifest bool) *Charm {
+func addCustomCharmWithManifest(c *gc.C, st *State, repo *repo.CharmRepo, name, filename, content, series string, revision int, manifest bool) *Charm {
 	path := repo.ClonedDirPath(c.MkDir(), name)
 	if filename != "" {
 		if manifest {
@@ -408,7 +409,7 @@ bases:
 	return addCharm(c, st, series, ch)
 }
 
-func addCustomCharm(c *gc.C, st *State, repo *testcharms.CharmRepo, name, filename, content, series string, revision int) *Charm {
+func addCustomCharm(c *gc.C, st *State, repo *repo.CharmRepo, name, filename, content, series string, revision int) *Charm {
 	return addCustomCharmWithManifest(c, st, repo, name, filename, content, series, revision, false)
 }
